+++ src/components/Collapse/Panel/style/Panel.style.ts
@@ -1,4 +1,5 @@
 import { styled } from '../../../../styled';
+import { StyledIcon } from '../../../Icon';
 
 export const StyledPanel = styled.div``;
 export const StyledPanelContent = styled.div``;
@@ -6,4 +7,11 @@ export const StyledPanelHeader = styled.div`
     display: flex;
     flex-direction: row;
     align-items: center;
+    cursor: pointer;
+
+    &:hover {
+        ${StyledIcon} {
+            color: ${(props) => props.theme.color.primary.alpha};
+        }
+    }
 `;