+++ src/components/Message/types/StyledMessageProps.ts
@@ -1,0 +1,6 @@
+import { CommonHTMLProps, MouseEvents, Type } from '../../../types';
+
+export type StyledMessageProps = CommonHTMLProps<HTMLDivElement> &
+    MouseEvents<HTMLDivElement> & {
+        type?: Type;
+    };