+++ src/index.ts
@@ -5,3 +5,4 @@ export * from './context';
 export * from './types/utils';
 export * from './styled';
 export * from './utils/wrap';
+export * from './utils/excludeProps';