+++ src/types/utils/InnerRef.ts
@@ -1,0 +1,3 @@
+export type InnerRef<T extends { ref?: any }> = Omit<T, 'ref'> & {
+    innerRef?: T['ref'];
+};